created simple mode from No_Auth tag
